requirements.txt
grpcio==1.60.0
grpcio-tools==1.60.0
protobuf==6.31.1